+++ src/main/java/net/berserker_rpg/custom/custom_spells/BerserkerSpellSchool.java
@@ -1,5 +1,6 @@
 package net.berserker_rpg.custom.custom_spells;
 
+import net.berserker_rpg.BerserkerClassMod;
 import net.minecraft.enchantment.EnchantmentHelper;
 import net.minecraft.enchantment.Enchantments;
 import net.minecraft.entity.attribute.EntityAttributes;
@@ -14,7 +15,7 @@ import net.spell_power.api.SpellSchools;
 import static net.spell_power.api.SpellPowerMechanics.PERCENT_ATTRIBUTE_BASELINE;
 
 public class BerserkerSpellSchool {
-    public static final SpellSchool BERSERKER_MELEE = new SpellSchool(SpellSchool.Archetype.MAGIC,
+    public static final SpellSchool BERSERKER_MELEE = new SpellSchool(SpellSchool.Archetype.MELEE,
             Identifier.of(SpellPowerMod.ID, "berserker_melee"),
             0xb3b3b3,
             DamageTypes.PLAYER_ATTACK,
@@ -22,7 +23,8 @@ public class BerserkerSpellSchool {
 
     public static void initialize() {
         BERSERKER_MELEE.addSource(SpellSchool.Trait.POWER, SpellSchool.Apply.ADD, query -> {
-            var power = query.entity().getAttributeValue(EntityAttributes.GENERIC_ATTACK_DAMAGE);
+            var power = query.entity().getAttributeValue(EntityAttributes.GENERIC_ATTACK_DAMAGE) +
+                    ((query.entity().getAttributeValue(MRPGCEntityAttributes.RAGE_MODIFIER)-100) / 50) ;
             var world = query.entity().getWorld();
             var sharpness = world.getRegistryManager().get(RegistryKeys.ENCHANTMENT).getEntry(Enchantments.SHARPNESS);
             if (sharpness.isPresent()) {
@@ -33,15 +35,13 @@ public class BerserkerSpellSchool {
         });
         BERSERKER_MELEE.addSource(SpellSchool.Trait.CRIT_CHANCE, new SpellSchool.Source(SpellSchool.Apply.ADD, query ->  {
             var value = SpellPowerMod.attributesConfig.value.base_spell_critical_chance_percentage
-                    + query.entity().getAttributeValue(MRPGCEntityAttributes.RAGE_MODIFIER);
-            var rate = (value / PERCENT_ATTRIBUTE_BASELINE);
-            return (rate - 1)/10;
+                    + query.entity().getAttributeValue(MRPGCEntityAttributes.RAGE_MODIFIER )- 100/ 10;
+            return (value/ PERCENT_ATTRIBUTE_BASELINE)-1;
         }));
         BERSERKER_MELEE.addSource(SpellSchool.Trait.CRIT_DAMAGE, new SpellSchool.Source(SpellSchool.Apply.ADD, query -> {
             var value = SpellPowerMod.attributesConfig.value.base_spell_critical_damage_percentage
-                    + query.entity().getAttributeValue(MRPGCEntityAttributes.RAGE_MODIFIER);
-            var rate = (value / PERCENT_ATTRIBUTE_BASELINE);
-            return (rate - 1)/4;
+                    + query.entity().getAttributeValue(MRPGCEntityAttributes.RAGE_MODIFIER )- 100/ 4;
+            return (value/ PERCENT_ATTRIBUTE_BASELINE)-1;
         }));
         SpellSchools.configureSpellCritDamage(BERSERKER_MELEE);
         SpellSchools.configureSpellCritChance(BERSERKER_MELEE);