+++ src/main/java/net/berserker_rpg/effect/BloodSacrificeEffect.java
@@ -1,0 +1,28 @@
+package net.berserker_rpg.effect;
+
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.effect.StatusEffect;
+import net.minecraft.entity.effect.StatusEffectCategory;
+
+public class BloodSacrificeEffect extends StatusEffect {
+    private int healthPerStack;
+
+    public BloodSacrificeEffect(StatusEffectCategory category, int color) {
+        super(category, color);
+        this.healthPerStack = 2  ;
+    }
+
+    public boolean applyUpdateEffect(LivingEntity entity, int amplifier) {
+        return entity.getAbsorptionAmount() > 0.0F || entity.getWorld().isClient;
+    }
+
+    public boolean canApplyUpdateEffect(int duration, int amplifier) {
+        return true;
+    }
+
+    public void onApplied(LivingEntity entity, int amplifier) {
+        super.onApplied(entity, amplifier);
+        entity.setAbsorptionAmount(Math.max(entity.getAbsorptionAmount(), healthPerStack  * (1 + amplifier )));
+    }
+
+}