+++ src/main/java/net/berserker_rpg/effect/Effects.java
@@ -38,6 +38,7 @@ public class Effects {
     public static final Entry RAGE=  new Entry("rage",new CustomStatusEffect(StatusEffectCategory.BENEFICIAL, 0xf70000));
     public static final Entry SOUL_DEVOURER =  new Entry("soul_devourer",new CustomStatusEffect(StatusEffectCategory.BENEFICIAL, 0x01d9cf));
     public static final Entry BLOOD_RECKONING=  new Entry("blood_reckoning",new BloodReckoningEffect(StatusEffectCategory.BENEFICIAL, 0xf70000));
+    public static final Entry BLOOD_SACRIFICE =  new Entry("blood_sacrifice",new BloodSacrificeEffect(StatusEffectCategory.BENEFICIAL, 0xf70000));
 
 
     public static void register(){
@@ -46,10 +47,13 @@ public class Effects {
                         effectsConfig.value.rage_rage_attribute_increase_per_stack, EntityAttributeModifier.Operation.ADD_MULTIPLIED_BASE)
                 .addAttributeModifier(EntityAttributes.GENERIC_ATTACK_SPEED, RAGE.modifierId(),
                         effectsConfig.value.rage_attack_speed_increase_per_stack, EntityAttributeModifier.Operation.ADD_MULTIPLIED_BASE);
+        BLOOD_SACRIFICE.effect.addAttributeModifier(EntityAttributes.GENERIC_MAX_ABSORPTION,
+                BLOOD_SACRIFICE.modifierId(), 4, EntityAttributeModifier.Operation.ADD_VALUE);
 
         Synchronized.configure(RAGE.effect,true);
         Synchronized.configure(SOUL_DEVOURER.effect,true);
         Synchronized.configure(BLOOD_RECKONING.effect,true);
+        Synchronized.configure(BLOOD_SACRIFICE.effect,true);
 
         HealthImpacting.configureDamageTaken(RAGE.effect,effectsConfig.value.rage_increased_incoming_damage_per_stack);